Apply int64 everywhere in merge_components; try not the let numba guess

+++ fast_hdbscan/boruvka.py
@@ -10,12 +10,12 @@ def merge_components(disjoint_set, candidate_neighbors, candidate_neighbor_dista
 
     # Find the best edges from each component
     for i in range(candidate_neighbors.shape[0]):
-        from_component = point_components[i]
+        from_component = np.int64(point_components[i])
         if from_component in component_edges:
             if candidate_neighbor_distances[i] < component_edges[from_component][2]:
-                component_edges[from_component] = (i, candidate_neighbors[i], candidate_neighbor_distances[i])
+                component_edges[from_component] = (np.int64(i), np.int64(candidate_neighbors[i]), candidate_neighbor_distances[i])
         else:
-            component_edges[from_component] = (i, candidate_neighbors[i], candidate_neighbor_distances[i])
+            component_edges[from_component] = (np.int64(i), np.int64(candidate_neighbors[i]), candidate_neighbor_distances[i])
 
     result = np.empty((len(component_edges), 3), dtype=np.float64)
     result_idx = 0
